[ISSUE #5181]♻️Refactor authentication context modules and update license headers
