+++ rocketmq-auth/src/authorization/context.rs
@@ -16,7 +16,5 @@
 //  under the License.
 
 pub mod authentication_context;
-pub mod base_authentication_context;
 pub mod base_authorization_context;
-pub mod default_authentication_context;
 pub mod default_authorization_context;
